+++ mcpackage/data/templates/python/__init__.py
@@ -1,0 +1,7 @@
+# coding: utf-8
+"""
+This package provides access to the python classes implemented in the
+sub-modules.
+"""
+
+from ${mod_namespace}.py.mod import ${mod_class}