+++ CHANGES.rst
@@ -2,6 +2,16 @@
 Change History
 ==============
 
+0.3.0 (TBD)
+-----------
+
+- Refactored main script to allow commands.
+- Added init command.
+- TODO: Refactored build command.
+- TODO: Added install command.
+- TODO: Added run command.
+
+
 0.2.0 (2013-12-18)
 ------------------